requirements.txt
python-dotenv==0.14.0
mysql-connector-python==8.3.0
playwright==1.41.2
urllib3==1.26.18
